First commit, contains all working files
First commit, contains all working files including databases which will
need to be inserted.

Init and InitPH will also need to be used in conjunction with reset.php.

NOTE: This program assumes the use of Shibboleth as its login management
client. Other forms of login will have to be programmed in a different
manner.